restructuring & CMAKE
- modernize CMAKE files
    * rename Test Targets
    * rename namespace and package
- moving code around
    * header files now reside in the src folder too
- stop calling std::stoN() in case the field is NULL to avoid unnecessary exceptions
- drop make_unique.h
